requirements.txt
fastapi==0.104.1
langchain==0.0.333
python-dotenv==1.0.0
streamlit==1.28.1
streamlit_chat==0.1.1
uvicorn==0.24.0.post1
azure-search-documents==11.4.0b8
azure-identity==1.15.0
tiktoken==0.5.1
openai==0.28.1
azure-ai-formrecognizer==3.3.2